feat: migrate basic dashboard to advanced dashboard
BREAKING CHANGE: Remove basic dashboard and consolidate all functionality into advanced dashboard

- Remove basic dashboard route (/dashboard) and replace with redirect to /advanced-dashboard
- Update header navigation to make advanced dashboard the primary dashboard
- Migrate missing 'Alerts & Notifications' widget to advanced dashboard
- Add comprehensive widget management features (category filtering, widget toggle)
- Sync all widgets between advanced-dashboard and advanced-dashboard-dnd
- Update documentation to reflect single advanced dashboard approach
- Ensure all 15 widgets are available in both advanced dashboard versions

Features added to advanced dashboard:
- Category filtering (All, Real-time, Monitor, Analytics, Controls)
- Widget toggle functionality
- Proper alerts widget integration
- Consistent widget set across all dashboard versions

Migration benefits:
- Single, unified dashboard experience
- All functionality preserved and enhanced
- Better user experience with advanced features
- Simplified maintenance and development
